added express validation for IDs

+++ server/controllers/employees.js
@@ -32,6 +32,10 @@ const createEmployee = async (req, res) => {
 }
 
 const deleteEmployee = async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ code: 400, error: errors.array()[0].msg });
+  }
   try {
     await Task.deleteMany({ "employeeId": req.params.id });
     await Employee.findByIdAndRemove(req.params.id);
